requirements.txt
streamlit==1.36.0
python-dotenv==1.0.1
pypdf==4.2.0
pdf2image==1.17.0
pytesseract==0.3.10
langchain==0.2.7
langchain-openai==0.1.14
langchain-community==0.2.6
langchain-mongodb==0.1.6
pymongo==4.8.0
